backend/requirements.txt
Django==5.0.3
asgiref==3.7.2
sqlparse==0.4.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.5
django-taggit==5.0.1
uuid==1.30
drf-extra-fields==3.7.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
